riscv: implmeent ThreadIdProvider for chips
