+++ chips/arty_e21_chip/src/chip.rs
@@ -146,6 +146,7 @@ impl<'a, I: InterruptService + 'a> ArtyExx<'a, I> {
 impl<'a, I: InterruptService + 'a> kernel::platform::chip::Chip for ArtyExx<'a, I> {
     type MPU = PMPUserMPU<2, SimplePMP<4>>;
     type UserspaceKernelBoundary = rv32i::syscall::SysCall;
+    type ThreadIdProvider = rv32i::thread_id::RiscvThreadIdProvider;
 
     fn mpu(&self) -> &Self::MPU {
         &self.pmp
@@ -231,3 +232,11 @@ pub extern "C" fn disable_interrupt_trap_handler(mcause: u32) {
         rv32i::clic::disable_interrupt(interrupt_index);
     }
 }
+
+/// Array used to track the "trap handler active" state per hart.
+///
+/// The `riscv` crate requires chip crates to allocate an array to
+/// track whether any given hart is currently in a trap handler. The
+/// array must be zero-initialized.
+#[export_name = "_trap_handler_active"]
+static mut TRAP_HANDLER_ACTIVE: [usize; 1] = [0; 1];
