+++ chips/qemu_rv32_virt_chip/src/chip.rs
@@ -114,7 +114,7 @@ impl<'a, I: InterruptService + 'a> QemuRv32VirtChip<'a, I> {
 impl<'a, I: InterruptService + 'a> Chip for QemuRv32VirtChip<'a, I> {
     type MPU = QemuRv32VirtPMP;
     type UserspaceKernelBoundary = rv32i::syscall::SysCall;
-    type ThreadIdProvider = rv32i::support::RiscvThreadIdProvider;
+    type ThreadIdProvider = rv32i::thread_id::RiscvThreadIdProvider;
 
     fn mpu(&self) -> &Self::MPU {
         &self.pmp
